MAT-419: Move unconverted skipped tests from integration back to unit
